+++ lib/init.c
@@ -138,7 +138,7 @@ void rpc_free_all_fragments(struct rpc_context *rpc)
 	}
 }
 
-int rpc_add_fragment(struct rpc_context *rpc, char *data, size_t size)
+int rpc_add_fragment(struct rpc_context *rpc, char *data, uint64_t size)
 {
 	struct rpc_fragment *fragment;
